updated articles sync 3;

+++ app/Services/ImageService.php
@@ -58,14 +58,28 @@ class ImageService {
             // Procesăm imaginea
             $image = ImageManager::read($file);
             $image->scaleDown(1000);
-            $image->save(storage_path('app/public/images/alpha/' . $articleImage->ImageFileName));
 
-            $optimizedImage = Storage::disk('public')->get('images/alpha/' . $articleImage->ImageFileName);
+            $optimizedImagePath = storage_path('app/public/images/alpha/' . $articleImage->ImageFileName);
+            $image->save($optimizedImagePath);
+
+            // Verificăm dacă imaginea optimizată a fost salvată corect
+            if (!file_exists($optimizedImagePath)) {
+                throw new \Exception('Fișierul optimizat nu a fost generat: ' . $articleImage->ImageFileName);
+            }
+
+            // Citim imaginea optimizată pentru încărcare
+            $optimizedImage = file_get_contents($optimizedImagePath);
+            if ($optimizedImage === false) {
+                throw new \Exception('Nu s-a putut citi fișierul optimizat: ' . $articleImage->ImageFileName);
+            }
+
+            // Încărcăm imaginea pe serverul SFTP
             $sshDisk = Storage::disk('sftp');
             $sshDisk->put($articleImage->ImageFileName, $optimizedImage);
 
             // Curățăm fișierele temporare
             Storage::disk('local')->delete('temp/' . $articleImage->ImageFileName);
+            Storage::disk('public')->delete('images/alpha/' . $articleImage->ImageFileName);
 
             return [
                 'width' => $image->width(),
@@ -78,7 +92,10 @@ class ImageService {
             dump($exception->getFile());
         }
 
-        Storage::disk('public')->delete('images/alpha/' . $articleImage->ImageFileName);
+        // Ștergem imaginea optimizată dacă s-a generat dar a apărut o eroare
+        if (isset($optimizedImagePath) && file_exists($optimizedImagePath)) {
+            unlink($optimizedImagePath);
+        }
     }
 
 }
